+++ lib/gen/dartling_model_generic.dart
@@ -14,8 +14,9 @@ String genEntries(Model model, String library) {
   sc = '${sc} \n';
   sc = '${sc}  Map<String, Entities> newEntries() { \n';
   sc = '${sc}    var entries = new Map<String, Entities>(); \n';
+  sc = '${sc}    var concept; \n';
   for (Concept entryConcept in model.entryConcepts) {
-    sc = '${sc}    var concept = model.concepts.singleWhereCode('
+    sc = '${sc}    concept = model.concepts.singleWhereCode('
          '"${entryConcept.code}"); \n';
     sc = '${sc}    entries["${entryConcept.code}"] = '
          'new ${entryConcept.codes}(concept); \n';