+++ lib/meta/types.dart
@@ -6,14 +6,20 @@ class AttributeTypes extends Entities<AttributeType> {
 
 class AttributeType extends ConceptEntity<AttributeType> {
 
-  String base;
+  String origin; 
   int length;
 
   Domain domain;
 
   AttributeType(this.domain, String typeCode) {
-    super.code = typeCode;
-    base = typeCode;
+    if (typeCode == 'Email') {
+      code = 'String';
+    } else if (typeCode == 'Other') {
+      code = 'dynamic';
+    } else {
+      code = typeCode;
+    }
+    origin = typeCode;
     domain.types.add(this);
   }